+++ src/npc.cpp
@@ -1609,8 +1609,8 @@ int npc::value(const item &it)
         auto &book = *it.type->book;
         if( book.intel <= int_cur ) {
             ret += book.fun;
-            if( book.skill != nullptr && skillLevel( book.skill ) < static_cast<int>( book.level ) &&
-                skillLevel( book.skill ) >= static_cast<int>( book.req ) ) {
+            if( book.skill != nullptr && skillLevel( book.skill ) < book.level &&
+                skillLevel( book.skill ) >= book.req ) {
                 ret += book.level * 3;
             }
         }
